+++ .mocharc.js
@@ -3,7 +3,7 @@
 module.exports = {
   delay: true,
   exclude: 'test/*/**',
-  reporter: 'cypress-multi-reporters',
+  reporter: 'mocha-multi-reporters',
   'reporter-option': [
     'configFile=mocha-multi-reporters.json'
   ],