+++ CHANGES.md
@@ -2,9 +2,12 @@
 
 ## ?
 
-- npm: Update devDeps (mocha-badge-generator)
-- npm: Use fork of cypress-multi-reporters (over fork of MMR) in
+- Docs: Update badges per latest updates
+- npm: Update from deprecated `rollup-plugin-babel` to `@rollup/plugin-babel`
+    (and make `babelHelpers` explicit)
+- npm: Use fork of `cypress-multi-reporters` (over fork of MMR) in
   preparation for using maintained fork
+- npm: Update devDeps
 
 ## 0.9.0